+++ Network/Packets/S_RoomReadyPacket.cs
@@ -5,11 +5,11 @@ namespace block_racing_common.Network.Packets
     {
         public PacketId PacketId => PacketId.S_RoomReady;
 
-        public int RoomId { get; set; }
+        public long RoomId { get; set; }
 
         public void Read(PacketReader reader)
         {
-            RoomId = reader.ReadInt32();
+            RoomId = reader.ReadLong();
         }
 
         public void Write(PacketWriter writer)
